+++ src/config.rs
@@ -1,6 +1,6 @@
 use directories::ProjectDirs;
 use serde::Deserialize;
-use std::fs;
+use std::{fs, path::PathBuf};
 use thiserror::Error;
 
 #[derive(Debug, Deserialize)]
@@ -47,25 +47,30 @@ pub enum ConfigError {
     },
 }
 
-pub fn load_config() -> Result<Config, ConfigError> {
+fn default_path() -> Result<PathBuf, ConfigError> {
     match ProjectDirs::from("com", "jonahmakowski", "cli-tool") {
-        Some(proj_dirs) => {
-            let config_path = proj_dirs.config_dir().join("config.yaml");
-
-            match fs::read_to_string(&config_path) {
-                Ok(config_text) => match yaml_serde::from_str::<Config>(&config_text) {
-                    Ok(config) => Ok(config),
-                    Err(err) => Err(ConfigError::Parse {
-                        path: config_path,
-                        source: err,
-                    }),
-                },
-                Err(err) => Err(ConfigError::Read {
-                    path: config_path,
-                    source: err,
-                }),
-            }
-        }
+        Some(proj_dirs) => Ok(proj_dirs.config_dir().join("config.yaml")),
         None => Err(ConfigError::NoProjectDirectory),
+    }
+}
+
+pub fn load_config(path: Option<PathBuf>) -> Result<Config, ConfigError> {
+    let config_path = match path {
+        Some(p) => p,
+        None => default_path()?,
+    };
+
+    match fs::read_to_string(&config_path) {
+        Ok(config_text) => match yaml_serde::from_str::<Config>(&config_text) {
+            Ok(config) => Ok(config),
+            Err(err) => Err(ConfigError::Parse {
+                path: config_path,
+                source: err,
+            }),
+        },
+        Err(err) => Err(ConfigError::Read {
+            path: config_path,
+            source: err,
+        }),
     }
 }