+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Moved config into lazylock
+- Replace string formatting with formatdoc for yaml config
 
 ### Fixed