+++ CHANGELOG.md
@@ -5,6 +5,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Made releases work
+
+## [0.6.0] - 2026-08-03
+
 ### Added
 
 - Add doctor command for system health checks
@@ -198,7 +204,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Update .forgejo/workflows/build.yaml
 - Update .forgejo/workflows/build.yaml
 
-[unreleased]: https://github.com/jonahmakowski/cli-tool/compare/v0.5.2..HEAD
+[unreleased]: https://github.com/jonahmakowski/cli-tool/compare/v0.6.0..HEAD
+[0.6.0]: https://github.com/jonahmakowski/cli-tool/compare/v0.5.2..v0.6.0
 [0.5.2]: https://github.com/jonahmakowski/cli-tool/compare/v0.5.1..v0.5.2
 [0.5.1]: https://github.com/jonahmakowski/cli-tool/compare/v0.5.0..v0.5.1
 [0.5.0]: https://github.com/jonahmakowski/cli-tool/compare/v0.4.0..v0.5.0